Develop

+++ src/main/java/com/FelipeMovio/banco/dto/DadosDto.java
@@ -2,10 +2,12 @@ package com.FelipeMovio.banco.dto;
 
 import com.FelipeMovio.banco.database.model.DadosEntity;
 
+import java.math.BigDecimal;
+
 public record DadosDto(
         Long numero,
         String agencia,
-        Double saldo
+        BigDecimal saldo
 ) {
     public DadosDto(DadosEntity dados) {
         this(dados.getNumero(),
